requirements.txt
aiogram==3.3.0
requests==2.31.0
beautifulsoup4==4.12.2
fake-useragent==1.4.0
aiohttp==3.9.0
python-dotenv==1.0.0